Dynamic model of the arm for tensions in Matlab
Were develped three aditional scripts in Matlab to analyze the tension
in the cable for the elbow movement.
Tension calc calculates the tension dynamic with variation of the elbow
angle and variation in the distance of the forearm clamping,
tension_dynamic takes care of the variation of the distance in the arm
clamping and finally the ploting_arm function graphs the arm and force
vs dcf.